requirements.txt
altair==5.1.2
asttokens==2.0.5
attrs==22.2.0
azure-ai-ml==1.5.0
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-storage-blob==12.15.0
azure-storage-file-datalake==12.10.1
azure-storage-file-share==12.11.1
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.2.0
colorama==0.4.4
commonmark==0.9.1
cryptography==40.0.1
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fugashi==1.1.2
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.11.0
google-auth==2.17.1
googleapis-common-protos==1.59.0
idna==3.3
importlib-metadata==6.8.0
ipykernel==6.13.0
ipython==8.3.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jisho-api==0.1.8
joblib==1.1.0
jsonschema==4.17.3
jupyter-client==7.3.0
jupyter-core==4.10.0
llvmlite==0.39.1
lxml==4.8.0
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib-inline==0.1.3
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
nest-asyncio==1.5.5
nltk==3.2.4
numba==0.56.2
numpy==1.22.3
oauthlib==3.2.2
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
openpyxl==3.1.2
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
plotly==5.17.0
portalocker==2.7.0
preprocessing==0.1.13
prompt-toolkit==3.0.29
protobuf==4.22.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pydash==5.1.2
pydeck==0.8.1b0
Pygments==2.11.2
PyJWT==2.6.0
pyparsing==3.0.8
pyrsistent==0.19.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
PyYAML==6.0
pyzmq==22.3.0
regex==2023.10.3
requests==2.27.1
requests-oauthlib==1.3.1
rich==10.16.2
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
shap==0.41.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
slugify==0.0.1
smart-open==6.3.0
smmap==5.0.1
soupsieve==2.3.2.post1
sphinx-rtd-theme==0.2.4
stack-data==0.2.0
streamlit==1.27.2
streamlit-shap==1.0.2
strictyaml==1.7.3
tenacity==8.2.3
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
underthesea==6.8.0
underthesea_core==1.0.4
unidic-lite==1.0.8
urllib3==1.26.9
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.5
zipp==3.17.0
